% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aperol.R
\name{aperol}
\alias{aperol}
\title{aperol}
\usage{
aperol(quant, ...)
}
\arguments{
\item{quant}{Wie viele sollen es werden?}
}
\value{
Wie Du einen Aperol Spritz machst.
}
\description{
Ein simples Rezept für Aperol Spritz mitten in R. Ist es nicht wunderbar.
}
