% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negroni.R
\name{negroni}
\alias{negroni}
\title{negroni}
\usage{
negroni()
}
\value{
Hast du die Idee etwa von Instagram oder bist du alleine drauf gekommen..?
}
\description{
Uhh, sehr hip. Aber auch sehr lecker und geeignet für einen stilvollen Abend, stilvolle Gespräche und das bisschen Philosophie.
}
