% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wassollichtun.R
\name{wassollichtun}
\alias{wassollichtun}
\title{wassollichtun}
\usage{
wassollichtun()
}
\value{
Ist nur ein Vorschlag,
}
\description{
Manchmal gibt es diese Zeiten am Tag, da muss man einfach R fragen...
}
